Move EarlyAdoptersListItem component to seperate file
